requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pymupdf==1.23.8
google-genai
pydantic==2.7.4
pillow==10.3.0
streamlit
